+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - bugfix [#1305](https://github.com/ergonode/backend/issues/1305) Adding validation for editing binding attributes in variable product.(wiewiurdp)
 - feature [#1283](https://github.com/ergonode/backend/issues/1283) Adding endpoint for status in batch actions  (wiewiurdp)
+- feature [#1318](https://github.com/ergonode/backend/issues/1318) Remove legacy Transformer Aggregate class (rprzedzik) 
 - refactor [#1271](https://github.com/ergonode/backend/issues/1271) Changed batch actions filtered endpoints to POST (piotrkreft)
 - refactor [#1272](https://github.com/ergonode/backend/issues/1272) Refactor completeness calculation process (rprzedzik)
 - bugfix [#1268](https://github.com/ergonode/backend/issues/1268) Export Shopware6 Configuration Segment list  (wfajczyk)